#12008 [fix] 년도별 지표관리, 년도별 sla현황 페이징처리 취소 및 기타작업

See merge request brainz.itsm/itsm!2736

+++ src/main/kotlin/co/brainz/itsm/sla/metricYear/dto/MetricYearDetailDto.kt
@@ -17,9 +17,9 @@ data class MetricYearDetailDto(
     var metricType: String,
     var metricUnit: String,
     var calculationType: String,
-    var minValue: Double,
-    var maxValue: Double,
-    var weightValue: Double,
+    var minValue: Float,
+    var maxValue: Float,
+    var weightValue: Float,
     var owner: String? = null,
     var comment: String? = null,
     var zqlString: String
